requirements.txt
Flask==3.0.2
Flask-MySQLdb==2.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
pandas==2.2.1
scikit-learn==1.4.1.post1
numpy==1.26.4
python-dotenv==1.0.1
mysql-connector-python==8.3.0
